Move exceptions into their own header
